requirements.txt
Flask
Flask-SQLAlchemy
requests
Pillow
celery
redis
pandas